# openapi/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Laravel Boilerplate API",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://laravel-boilerplate.test/api"
        }
    ],
...
    "security": [
        {
            "http": []
        }
    ],
    "paths": {
...
        "/forgot-password": {
            "post": {
                "operationId": "auth.forgotPassword",
                "description": "Sends a one-time passcode to the user's email to initiate the password reset flow.",
                "summary": "Forgot password",
                "tags": [
                    "Auth"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ForgetPassword"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "message"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
        "/forgot-password-otp-verify": {
            "post": {
                "operationId": "auth.forgotPasswordOTPVerify",
                "description": "Validates the OTP sent to the user's email and returns a password-reset token.",
                "summary": "Verify OTP for password reset",
                "tags": [
                    "Auth"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VerifyOtp"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "token": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "message",
                                        "token"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
        "/reset-password": {
            "post": {
                "operationId": "auth.resetPassword",
                "description": "Sets a new password using the reset token obtained after OTP verification.",
                "summary": "Reset password",
                "tags": [
                    "Auth"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ResetPassword"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "message"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
        "/me": {
            "get": {
                "operationId": "user.me",
                "description": "Returns the profile details of the currently authenticated user.",
                "summary": "Get authenticated user",
                "tags": [
                    "Auth"
                ],
                "responses": {
                    "200": {
                        "description": "`User`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            },
            "post": {
                "operationId": "user.updateProfile",
                "description": "Updates the profile information for the authenticated user.",
                "summary": "Update profile",
                "tags": [
                    "Auth"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProfile"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "required": [
                                        "message",
                                        "user"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
        "/change-password": {
            "post": {
                "operationId": "user.changePassword",
                "description": "Changes the authenticated user's password after verifying the current one.",
                "summary": "Change password",
                "tags": [
                    "Auth"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChangePassword"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "message"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "http": {
                "type": "http",
                "scheme": "bearer"
            }
        },
        "schemas": {
...
            "ChangePassword": {
                "type": "object",
                "properties": {
                    "current_password": {
                        "type": "string",
                        "minLength": 8
                    },
                    "password": {
                        "type": "string",
                        "minLength": 8
                    },
                    "password_confirmation": {
                        "type": "string",
                        "minLength": 8
                    }
                },
                "required": [
                    "current_password",
                    "password",
                    "password_confirmation"
                ],
                "title": "ChangePassword"
            },
...
            "ForgetPassword": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "maxLength": 255
                    }
                },
                "required": [
                    "email"
                ],
                "title": "ForgetPassword"
            },
...
            "Media": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "The unique identifier of the media file."
                    },
                    "disk": {
                        "type": "string",
                        "description": "The storage disk where the file is stored (e.g. local, s3)."
                    },
                    "directory": {
                        "type": "string",
                        "description": "The directory path within the disk."
                    },
                    "filename": {
                        "type": "string",
                        "description": "The filename without extension."
                    },
                    "extension": {
                        "type": "string",
                        "description": "The file extension (e.g. jpg, png)."
                    },
                    "mime_type": {
                        "type": "string",
                        "description": "The MIME type of the file."
                    },
                    "aggregate_type": {
                        "type": "string",
                        "description": "The aggregate type category (e.g. image, video, document)."
                    },
                    "size": {
                        "type": "integer",
                        "description": "The file size in bytes."
                    },
                    "url": {
                        "type": "string",
                        "format": "uri",
                        "description": "The public URL to access the file."
                    },
                    "cdn_url": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "uri",
                        "description": "The CDN URL for the file, if CDN is enabled."
                    }
                },
                "required": [
                    "id",
                    "disk",
                    "directory",
                    "filename",
                    "extension",
                    "mime_type",
                    "aggregate_type",
                    "size",
                    "url",
                    "cdn_url"
                ],
                "title": "Media"
            },
...
            "ResetPassword": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "maxLength": 255
                    },
                    "password": {
                        "type": "string",
                        "minLength": 8
                    },
                    "token": {
                        "type": "string"
                    },
                    "password_confirmation": {
                        "type": "string",
                        "minLength": 8
                    }
                },
                "required": [
                    "email",
                    "password",
                    "token",
                    "password_confirmation"
                ],
                "title": "ResetPassword"
            },
...
            "UpdateProfile": {
                "type": "object",
                "properties": {
                    "first_name": {
                        "type": "string",
                        "maxLength": 120
                    },
                    "last_name": {
                        "type": "string",
                        "maxLength": 120
                    },
                    "username": {
                        "type": "string",
                        "maxLength": 120
                    },
                    "country_code": {
                        "type": "string",
                        "maxLength": 5
                    },
                    "mobile_no": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "minLength": 8,
                        "maxLength": 15
                    },
                    "profile": {
                        "type": "object",
                        "properties": {
                            "filename": {
                                "type": "string",
                                "maxLength": 255
                            },
                            "directory": {
                                "type": "string",
                                "maxLength": 255
                            },
                            "size": {
                                "type": "integer"
                            },
                            "mime_type": {
                                "type": "string",
                                "enum": [
                                    "image/jpeg",
                                    "image/png",
                                    "image/gif",
                                    "image/webp"
                                ]
                            }
                        }
                    }
                },
                "required": [
                    "first_name",
                    "last_name",
                    "username",
                    "profile"
                ],
                "title": "UpdateProfile"
            },
            "User": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "The unique identifier of the user."
                    },
                    "first_name": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "description": "The user's first name."
                    },
                    "last_name": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "description": "The user's last name."
                    },
                    "name": {
                        "type": "string",
                        "description": "The full display name of the user."
                    },
                    "username": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "description": "The user's unique username."
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "description": "The user's email address."
                    },
                    "status": {
                        "description": "The user's account status.",
                        "$ref": "#/components/schemas/UserStatus"
                    },
                    "display_status": {
                        "type": "string",
                        "description": "Human-readable label for the account status."
                    },
                    "country_code": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "description": "The country dial code (e.g. +1)."
                    },
                    "mobile_no": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "description": "The user's mobile number."
                    },
                    "display_mobile_no": {
                        "type": "string",
                        "description": "Mobile number with country code prefix."
                    },
                    "email_verified_at": {
                        "type": [
                            "integer",
                            "null"
                        ],
                        "description": "Timestamp when the email was verified."
                    },
                    "last_login_at": {
                        "type": "integer",
                        "description": "Timestamp of the user's last login."
                    },
                    "created_at": {
                        "type": [
                            "integer",
                            "null"
                        ],
                        "description": "Timestamp when the user was created."
                    },
                    "profile_image": {
                        "$ref": "#/components/schemas/Media"
                    },
                    "user_device": {
                        "$ref": "#/components/schemas/UserDevice"
                    }
                },
                "required": [
                    "id",
                    "first_name",
                    "last_name",
                    "name",
                    "username",
                    "email",
                    "status",
                    "display_status",
                    "country_code",
                    "mobile_no",
                    "display_mobile_no",
                    "email_verified_at",
                    "last_login_at",
                    "created_at"
                ],
                "title": "User"
            },
            "UserDevice": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "The unique identifier of the device record."
                    },
                    "user_id": {
                        "type": [
                            "integer",
                            "null"
                        ],
                        "description": "The ID of the owning user."
                    },
                    "onesignal_player_id": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "description": "The OneSignal player ID for push notifications."
                    },
                    "device_id": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "description": "The unique hardware device ID."
                    },
                    "device_type": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "description": "The type of device (e.g. ios, android, web)."
                    }
                },
                "required": [
                    "id",
                    "user_id",
                    "onesignal_player_id",
                    "device_id",
                    "device_type"
                ],
                "title": "UserDevice"
            },
            "UserStatus": {
                "type": "string",
                "enum": [
                    "active",
                    "inactive"
                ],
                "title": "UserStatus"
            },
            "VerifyOtp": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "maxLength": 120
                    },
                    "otp": {
                        "type": "string"
                    }
                },
                "required": [
                    "email",
                    "otp"
                ],
                "title": "VerifyOtp"
            }
        },
        "responses": {
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            },
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
...
        }
    }
}